fix(maze): fix stutter (#60)
* fix(maze): fix stutter

- limit frame rate

* fix(maze): accept non monotonically increase

- SDL_GetPerformanceCounter() may not return monotonically increasing values. This happens in VMs and Emscripten.

* fix(maze): make const

* fix(maze): make physics timer a global

+++ game/src/exitlayer.h
@@ -10,7 +10,7 @@ class ExitLayer : public sponge::layer::Layer {
     void onAttach() override;
     void onDetach() override;
     void onEvent(sponge::event::Event& event) override;
-    bool onUpdate(uint32_t elapsedTime, bool isEventHandled) override;
+    bool onUpdate(double elapsedTime, bool isEventHandled) override;
 
     void setWidthAndHeight(uint32_t width, uint32_t height) const;
